pyproject.toml
[project]
name = "trun"
version = "0.1.0"
description = "Multi-Agent collaboration framework for local development"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "TeamRun Contributors" }
]
keywords = ["ai", "agent", "collaboration", "automation", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
# Agent SDKs as optional since they may not be available on PyPI yet
agents = [
    "claude-agent-sdk>=0.1.0",
    "codex-local-sdk-python>=0.1.0",
]

[project.scripts]
trun = "trun.cli:main"

[project.urls]
Homepage = "https://github.com/teamrun/teamrun"
Repository = "https://github.com/teamrun/teamrun"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["trun"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
